naver-blog-agent/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP Clients
httpx>=0.26.0
aiohttp>=3.9.0
curl-cffi>=0.6.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
playwright>=1.41.0

# LLM
google-generativeai>=0.4.0

# Utils
tenacity>=8.2.0
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
